p2p, refactor: extract Network code to node/network
