+++ src/Makefile.am
@@ -224,6 +224,7 @@ BITCOIN_CORE_H = \
   node/miner.h \
   node/mini_miner.h \
   node/minisketchwrapper.h \
+  node/network.h \
   node/peerman_args.h \
   node/psbt.h \
   node/transaction.h \
@@ -422,6 +423,7 @@ libbitcoin_node_a_SOURCES = \
   node/miner.cpp \
   node/mini_miner.cpp \
   node/minisketchwrapper.cpp \
+  node/network.cpp \
   node/peerman_args.cpp \
   node/psbt.cpp \
   node/transaction.cpp \
