+++ src/init.cpp
@@ -50,6 +50,7 @@
 #include <node/mempool_args.h>
 #include <node/mempool_persist_args.h>
 #include <node/miner.h>
+#include <node/network.h>
 #include <node/peerman_args.h>
 #include <node/validation_cache_args.h>
 #include <policy/feerate.h>
