+++ src/netaddress.h
@@ -11,6 +11,7 @@
 
 #include <compat/compat.h>
 #include <crypto/siphash.h>
+#include <node/network.h>
 #include <prevector.h>
 #include <random.h>
 #include <serialize.h>
@@ -31,42 +32,6 @@
  * or with `SERIALIZE_TRANSACTION_NO_WITNESS`.
  */
 static constexpr int ADDRV2_FORMAT = 0x20000000;
-
-/**
- * A network type.
- * @note An address may belong to more than one network, for example `10.0.0.1`
- * belongs to both `NET_UNROUTABLE` and `NET_IPV4`.
- * Keep these sequential starting from 0 and `NET_MAX` as the last entry.
- * We have loops like `for (int i = 0; i < NET_MAX; ++i)` that expect to iterate
- * over all enum values and also `GetExtNetwork()` "extends" this enum by
- * introducing standalone constants starting from `NET_MAX`.
- */
-enum Network {
-    /// Addresses from these networks are not publicly routable on the global Internet.
-    NET_UNROUTABLE = 0,
-
-    /// IPv4
-    NET_IPV4,
-
-    /// IPv6
-    NET_IPV6,
-
-    /// TOR (v2 or v3)
-    NET_ONION,
-
-    /// I2P
-    NET_I2P,
-
-    /// CJDNS
-    NET_CJDNS,
-
-    /// A set of addresses that represent the hash of a string or FQDN. We use
-    /// them in AddrMan to keep track of which DNS seeds were used.
-    NET_INTERNAL,
-
-    /// Dummy value to indicate the number of NET_* constants.
-    NET_MAX,
-};
 
 /// Prefix of an IPv6 address when it contains an embedded IPv4 address.
 /// Used when (un)serializing addresses in ADDRv1 format (pre-BIP155).
