+++ src/netbase.cpp
@@ -7,6 +7,7 @@
 
 #include <compat/compat.h>
 #include <logging.h>
+#include <node/network.h>
 #include <sync.h>
 #include <tinyformat.h>
 #include <util/sock.h>
@@ -75,54 +76,6 @@ std::vector<CNetAddr> WrappedGetAddrInfo(const std::string& name, bool allow_loo
 }
 
 DNSLookupFn g_dns_lookup{WrappedGetAddrInfo};
-
-enum Network ParseNetwork(const std::string& net_in) {
-    std::string net = ToLower(net_in);
-    if (net == "ipv4") return NET_IPV4;
-    if (net == "ipv6") return NET_IPV6;
-    if (net == "onion") return NET_ONION;
-    if (net == "tor") {
-        LogPrintf("Warning: net name 'tor' is deprecated and will be removed in the future. You should use 'onion' instead.\n");
-        return NET_ONION;
-    }
-    if (net == "i2p") {
-        return NET_I2P;
-    }
-    if (net == "cjdns") {
-        return NET_CJDNS;
-    }
-    return NET_UNROUTABLE;
-}
-
-std::string GetNetworkName(enum Network net)
-{
-    switch (net) {
-    case NET_UNROUTABLE: return "not_publicly_routable";
-    case NET_IPV4: return "ipv4";
-    case NET_IPV6: return "ipv6";
-    case NET_ONION: return "onion";
-    case NET_I2P: return "i2p";
-    case NET_CJDNS: return "cjdns";
-    case NET_INTERNAL: return "internal";
-    case NET_MAX: assert(false);
-    } // no default case, so the compiler can warn about missing cases
-
-    assert(false);
-}
-
-std::vector<std::string> GetNetworkNames(bool append_unroutable)
-{
-    std::vector<std::string> names;
-    for (int n = 0; n < NET_MAX; ++n) {
-        const enum Network network{static_cast<Network>(n)};
-        if (network == NET_UNROUTABLE || network == NET_INTERNAL) continue;
-        names.emplace_back(GetNetworkName(network));
-    }
-    if (append_unroutable) {
-        names.emplace_back(GetNetworkName(NET_UNROUTABLE));
-    }
-    return names;
-}
 
 static std::vector<CNetAddr> LookupIntern(const std::string& name, unsigned int nMaxSolutions, bool fAllowLookup, DNSLookupFn dns_lookup_function)
 {
