+++ src/node/eviction.h
@@ -5,13 +5,14 @@
 #ifndef BITCOIN_NODE_EVICTION_H
 #define BITCOIN_NODE_EVICTION_H
 
-#include <node/connection_types.h>
-#include <net_permissions.h>
+#include <node/network.h>
 
 #include <chrono>
 #include <cstdint>
 #include <optional>
 #include <vector>
+
+enum class ConnectionType;
 
 typedef int64_t NodeId;
