+++ src/qt/clientmodel.cpp
@@ -17,6 +17,7 @@
 #include <interfaces/node.h>
 #include <net.h>
 #include <netbase.h>
+#include <node/network.h>
 #include <util/threadnames.h>
 #include <util/time.h>
 #include <validation.h>
